Refactor core/tests into submodules
